+++ fact.h
@@ -6,8 +6,8 @@
 #include <unistd.h>
 
 FILE *fileOpener(const char *);
-unsigned long dtSwitcher(const char *, unsigned long);
+unsigned long long dtSwitcher(const char *, unsigned long long);
 void fileFactorizer(FILE *, char *, size_t);
-void numFactorizer(unsigned long);
+void numFactorizer(unsigned long long);
 void onExit(char *, FILE *);
 #endif /*FACT_H*/
